[update] - update user serializers, models, views
