+++ backend/users/models/custom_user.py
@@ -1,0 +1,6 @@
+from django.db import models
+from django.contrib.auth.models import AbstractBaseUser
+
+
+class CustomUser(AbstractBaseUser):
+    email = models.EmailField(unique=True, null=True, blank=True)
