+++ backend/users/serializers/auth_serializer.py
@@ -1,0 +1,25 @@
+from django.contrib.auth import authenticate
+
+from rest_framework import serializers
+
+
+class AuthSerializer(serializers.Serializer):
+    username = serializers.CharField()
+    password = serializers.CharField(
+        style={"input_type": "password"}, trim_whitespace=False
+    )
+
+    def validate(self, attrs):
+        username = attrs.get("username")
+        password = attrs.get("password")
+
+        user = authenticate(
+            request=self.context.get("request"), username=username, password=password
+        )
+
+        if not user:
+            msg = "Unable to authenticate with provided credentials"
+            raise serializers.ValidationError(msg, code="authentication")
+
+        attrs["user"] = user
+        return
