+++ backend/users/signals.py
@@ -2,7 +2,9 @@ import random
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.contrib.auth.models import User
-from .models import Profile, Settings
+
+from .models.profile import Profile
+from .models.settings import Settings
 
 from api.models.model import Model
